app10_web-scraping/requirements.txt
python-dotenv==1.0.1
Requests==2.31.0
selectorlib==0.16.0
